Modifica del tipo di "lato 1" da double a Double, secondo le specifiche consigliate.
Eliminazione della cartella "Package_Repository".

Aggiunto shutdown della connessione al database una volta completate le operazioni per un errore scaturito dall'inserimento di una seconda figura in successione alla prima, la cui connessione rimaneva attiva provocando l'errore "Database already in use".

+++ src/main/java/it/intesys/codylab/business/FormeGeometricheService.java
@@ -33,7 +33,7 @@ public class FormeGeometricheService {
         logger.info("area: {}", formaGeometrica.area());
     }
 
-    public void save(String tipo, int lato1, Double lato2) {
+    public void save(String tipo, Double lato1, Double lato2) {
         formaGeometricaRepository.save(tipo, lato1, lato2);
     }
 }
